chaolife/requirements.txt
Django == 1.8.3
pymysql == 0.7.4
django-jsonfield
django-filter
djangorestframework == 3.3.3
djangorestframework-jwt
django-email
django-admin-bootstrapped
django-crispy-forms
django-grappelli
django-enumfields == 0.7.4
django-redis-cache
requests == 2.10.0
drf-nested-routers
django-celery
dynamic_rest
drf-enum-field
drf-extensions
django-model-utils
qiniu
django-guardian
django-debug-toolbar
celery == 3.1.18
pycrypto
django-qiniu-storage
Pillow
py-moneyed
django-money


 
